[web] discover: extract shared components for iac integrations
