+++ web/packages/teleport/src/Integrations/Enroll/Cloud/Aws/RegionsSection.tsx
@@ -25,9 +25,8 @@ import { Rule } from 'shared/components/Validation/rules';
 import { Regions as AwsRegion } from 'teleport/services/integrations';
 
 import { RegionMultiSelector } from '../RegionMultiSelector';
-import { CircleNumber } from './EnrollAws';
+import { CircleNumber, WildcardRegion } from '../Shared';
 import { awsRegionGroups } from './regions';
-import { WildcardRegion } from './types';
 
 type RegionsOrWildcard = WildcardRegion | AwsRegion[];
